requirements.txt
fastapi==0.75.0
line-bot-sdk==2.2.1
uvicorn==0.17.6
langchain
openai==0.27.4
pytz
python-dotenv
guidance
psycopg2-binary